backend/foodgram/requirements.txt
Django==4.1
Pillow==8.3.1
requests==2.26.0
djangorestframework==3.13.1
django-filter==22.1
gunicorn==20.1.0
sorl-thumbnail==12.7.0
djangorestframework-simplejwt==5.2.0
drf-extra-fields==3.4.0
django-colorfield==0.7.2
psycopg2-binary==2.8.6